+++ docs/source/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changed
+
+- `moldrug.utils.Individual` now is a hashable object.
+
 ## [1.0.2] - 2022.08.08
 
 ### Removed